changed bad shit and added stuff
