+++ app/components/CardFivePreview/CardFivePreview.js
@@ -4,7 +4,7 @@ import { View, Text, Image } from 'react-native';
 import styles from './styles';
 import EStyleSheet from 'react-native-extended-stylesheet';
 
-const CardFivePreview = ({ city, stateabb, zip, logo, color, position, website, buisname, phonenum, email, address, name }) => {
+const CardFivePreview = ({ city, stateabb, zip, logo, color, position, website, businame, phonenum, email, address, name }) => {
 
       return (
         <View style={styles.top}>
@@ -24,7 +24,7 @@ const CardFivePreview = ({ city, stateabb, zip, logo, color, position, website, 
                             <Text style={styles.address2}>{city} {stateabb} {zip}</Text>
                             <Text style={styles.phonenum}>{phonenum}</Text>
                             <Text style={styles.name}>{name}</Text>
-                            <Text style={styles.buisname}>{buisname}</Text>
+                            <Text style={styles.businame}>{businame}</Text>
                             <Text style={styles.title}>{position}</Text>
                     </View>
                 )}
