+++ app/components/CardFivePreview/styles.js
@@ -29,7 +29,7 @@ export default EStyleSheet.create({
       width: (width-20),
       height: ((width-20)*.57),
   },
-  buisname: {
+  businame: {
       color: 'white',
       fontSize: 14,
       position: 'absolute',
