+++ app/components/CardFourPreview/CardFourPreview.js
@@ -4,7 +4,7 @@ import { View, Text, Image } from 'react-native';
 import styles from './styles';
 import EStyleSheet from 'react-native-extended-stylesheet';
 
-const CardFourPreview = ({ city, stateabb, zip, logo, color, position, website, buisname, phonenum, email, address, name }) => {
+const CardFourPreview = ({ city, stateabb, zip, logo, color, position, website, businame, phonenum, email, address, name }) => {
 
       return (
         <View style={styles.top}>
@@ -25,7 +25,7 @@ const CardFourPreview = ({ city, stateabb, zip, logo, color, position, website, 
                             <Text style={styles.address2}>{city} {stateabb} {zip}</Text>
                             <Text style={styles.website}>{website}</Text>
                             <Text style={styles.phonenum}>{phonenum}</Text>
-                            <Text style={styles.buisname}>{buisname}</Text>
+                            <Text style={styles.businame}>{businame}</Text>
                     </View>
                 )}
             />
