+++ app/components/CardTwoPreview/CardTwoPreview.js
@@ -4,7 +4,7 @@ import { View, Text, Image } from 'react-native';
 import styles from './styles';
 import EStyleSheet from 'react-native-extended-stylesheet';
 
-const CardTwoPreview = ({ city, stateabb, zip, logo, color, position, website, buisname, phonenum, email, address, name }) => {
+const CardTwoPreview = ({ city, stateabb, zip, logo, color, position, website, businame, phonenum, email, address, name }) => {
 
     return (
         <View style={styles.top}>
@@ -22,7 +22,7 @@ const CardTwoPreview = ({ city, stateabb, zip, logo, color, position, website, b
                             <Text style={styles.address}>{address}</Text>
                             <Text style={styles.name}>{name}</Text>
                             <Text style={styles.phonenum}>{phonenum}</Text>
-                            <Text style={styles.buisname}>{buisname}</Text>
+                            <Text style={styles.businame}>{businame}</Text>
                             <Text style={styles.address2}>{city} {stateabb} {zip}</Text>
                             <Text style={styles.title}>{position}</Text>
                             <Text style={styles.website}>{website}</Text>
