+++ app/screens/ProfileScreen.js
@@ -29,7 +29,7 @@ class ProfileScreen extends Component {
     }
 
     componentWillMount() {
-        store.get('buiscards').then((value) => {
+        store.get('busicards').then((value) => {
             if (value!==null){
                 this.setState({cards: value});
                 this.forceUpdate()
@@ -39,7 +39,7 @@ class ProfileScreen extends Component {
 
     _onRefresh() {
         this.setState({refreshing: true});
-        store.get('buiscards').then((value) => {
+        store.get('busicards').then((value) => {
             if (value!==null){
                 this.setState({cards: value});
                 this.forceUpdate();
@@ -93,7 +93,7 @@ class ProfileScreen extends Component {
                             position={ref.item.position}
                             color={ref.item.color}
                             website={ref.item.website}
-                            buisname={ref.item.buisname}
+                            businame={ref.item.businame}
                             phonenum={ref.item.phonenum}
                             name={ref.item.name}
                             email={ref.item.email}
@@ -132,7 +132,7 @@ class ProfileScreen extends Component {
                             position={ref.item.position}
                             color={ref.item.color}
                             website={ref.item.website}
-                            buisname={ref.item.buisname}
+                            businame={ref.item.businame}
                             phonenum={ref.item.phonenum}
                             name={ref.item.name}
                             email={ref.item.email}
@@ -171,7 +171,7 @@ class ProfileScreen extends Component {
                             position={ref.item.position}
                             color={ref.item.color}
                             website={ref.item.website}
-                            buisname={ref.item.buisname}
+                            businame={ref.item.businame}
                             phonenum={ref.item.phonenum}
                             name={ref.item.name}
                             email={ref.item.email}
@@ -210,7 +210,7 @@ class ProfileScreen extends Component {
                             position={ref.item.position}
                             color={ref.item.color}
                             website={ref.item.website}
-                            buisname={ref.item.buisname}
+                            businame={ref.item.businame}
                             phonenum={ref.item.phonenum}
                             name={ref.item.name}
                             email={ref.item.email}
@@ -249,7 +249,7 @@ class ProfileScreen extends Component {
                             position={ref.item.position}
                             color={ref.item.color}
                             website={ref.item.website}
-                            buisname={ref.item.buisname}
+                            businame={ref.item.businame}
                             phonenum={ref.item.phonenum}
                             name={ref.item.name}
                             email={ref.item.email}
@@ -268,7 +268,7 @@ class ProfileScreen extends Component {
         arrayCopy = this.state.cards
         arrayCopy.splice(index, 1)
         setTimeout(() => {this.setState({cards: arrayCopy})}, 215)
-        AsyncStorage.setItem('buiscards', JSON.stringify(arrayCopy))
+        AsyncStorage.setItem('busicards', JSON.stringify(arrayCopy))
     }
 
     _editItem(ref) {
